refactor: change package name to JS, not zollinger
