+++ build.gradle.kts
@@ -1,5 +1,5 @@
 version = "0.0.1"
-group = "com.jonathanzollinger.justserve"
+group = "org.justserve.helpcenter"
 
 val kotlinVersion= project.properties["kotlinVersion"]
 repositories {
