+++ core/build.gradle.kts
@@ -8,7 +8,7 @@ plugins {
 }
 
 version = "0.0.1"
-group = "com.jonathanzollinger.justserve"
+group = "org.justserve.helpcenter"
 
 val kotlinVersion= project.properties["kotlinVersion"]
 repositories {
