+++ core/src/main/kotlin/justserve/helpcenter/Application.kt
@@ -1,4 +1,4 @@
-package jonathanzollinger
+package justserve.helpcenter
 
 import io.micronaut.runtime.Micronaut.run
 fun main(args: Array<String>) {
