+++ core/src/test/groovy/justserve/helpcenter/JustServeSpec.groovy
@@ -1,4 +1,4 @@
-package jonathanzollinger
+package justserve.helpcenter
 
 import io.micronaut.runtime.EmbeddedApplication
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
